requirements.txt
streamlit
google-api-python-client
pdfplumber
google-auth-oauthlib
PyPDF2
python-docx
openai
python-dotenv
